requirements.txt
appdirs==1.4.0
hurry.filesize==0.9
packaging==16.8
pyparsing==2.1.10
pyvmomi==6.5
requests==2.13.0
six==1.10.0
